As a user, I can add a Crop to echarts
- User can add a Crop, Crop attributes and Crop interventions
- User can select a Crop to add some attributes or interventions
- Crop is displayed using provided render.js function

For the moment it's not possible to edit the values, to be done later

There is bug somewhere, maybe in render.js the start and enddate is not
good. May be fixed later if we don't switch to startDate in a Crop.

Extracted js from the example to another file
Added a css file for more clarity